+++ contracts/user/IProfile.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.0;
 
 interface IProfile {
 
-  event ValidityChanged(uint validity);
+  event ValidityChanged(uint256 validity);
   event AccountsAssociated(address addr1, address addr2);
   event AccountsDissociated(address addr1, address addr2);
 
-  function changeValidity(uint validity) external;
+  function changeValidity(uint256 validity) external;
 
-  function associateAccount(address account, uint timestamp, bytes memory signature) external;
+  function associateAccount(address account, uint256 timestamp, bytes memory signature) external;
 
   function dissociateAccount(address account) external;
 
@@ -17,6 +17,6 @@ interface IProfile {
 
   function isMyAssociated(address addr) external view returns (bool);
 
-  function encodeForSignature(address addr1, address addr2, uint timestamp) external view returns (bytes32);
+  function encodeForSignature(address addr1, address addr2, uint256 timestamp) external view returns (bytes32);
 
 }
